+++ member-app-external-api/src/main/java/io/howstheairtoday/memberappexternalapi/service/dto/request/MemberLoginRequestDTO.java
@@ -1,4 +1,4 @@
-package io.howstheairtoday.memberappexternalapi.service.dto;
+package io.howstheairtoday.memberappexternalapi.service.dto.request;
 
 import java.util.Collection;
